Adjust extendedGlob to work on Windows
Changes the logic of extendedGlob to be able to handle Windows
filepaths. The existing logic was unable to deal with paths starting
with a volumename (C:\), as it expected a leading path separator, which
will not be the case for local filepaths on Windows.

In order for the unit test to compile, the xattr test had to be excluded
on Windows, as it depended on a variable in xattr.go, which is excluded
on Windows.